GCG-02: As a developer, I want a PostgreSQL schema so that data persists reliably. - delete users_review table
