+++ database/factories/ReviewFactory.php
@@ -2,6 +2,7 @@
 
 namespace Database\Factories;
 
+use App\Models\User;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
 /**
@@ -17,6 +18,8 @@ class ReviewFactory extends Factory
     public function definition(): array
     {
         return [
+            'reviewer_id' => User::inRandomOrder()->first()?->id ?? User::factory(),
+            'reviewee_id' => User::inRandomOrder()->first()?->id ?? User::factory(),
             'comment' => $this->faker->sentence($nbWords = 15, $variableNbWords = true),
             'date' => $this->faker->dateTimeBetween('-5 years', 'now'),
         ];
