+++ database/seeders/DatabaseSeeder.php
@@ -6,9 +6,6 @@ use Illuminate\Database\Seeder;
 
 class DatabaseSeeder extends Seeder
 {
-    /**
-     * Seed the application's database.
-     */
     public function run(): void
     {
         $this->call([
@@ -21,7 +18,6 @@ class DatabaseSeeder extends Seeder
         $this->call([
             UserSkillSeeder::class,
             UserBadgeSeeder::class,
-            UsersReviewSeeder::class,
         ]);
     }
 }
